Model prj name change to Obj
