+++ OpenYourTable.Infra/OpenYourTable.Infra.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\OpenYourTable.Model\OpenYourTable.Model.csproj" />
+    <ProjectReference Include="..\OpenYourTable.Obj\OpenYourTable.Obj.csproj" />
   </ItemGroup>
 
 </Project>
