+++ OpenYourTable.Obj/Enums/EnumDB.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel;
 
-namespace OpenYourTable.Model.Enums
+namespace OpenYourTable.Obj.Enums
 {
 	public enum DB_TYPE
 	{
