+++ OpenYourTable.Obj/Enums/EnumExcel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel;
 
-namespace OpenYourTable.Model.Enums
+namespace OpenYourTable.Obj.Enums
 {
 	public enum TABLE_CELL
 	{
